+++ CHANGES.rst
@@ -7,6 +7,9 @@
 
 - Add a month list view which looks like the print version of the month view.
 
+- Allow to delete a recurred event and put a new event with the same category
+  into its place. Before this fix the new event might not be displayed.
+
 
 2.1.1 (2017-04-09)
 ==================